requirements.txt
Flask
requests
gunicorn
PyQt6
psycopg2-binary
python-dotenv